WIP: refactor all grib-file-based inputs
